Skip `git clean` in reset_git_repo

+++ lib/fastlane/actions/reset_git_repo.rb
@@ -10,7 +10,7 @@ module Fastlane
 
           if (paths || []).count == 0
             Actions.sh('git reset --hard HEAD')
-            Actions.sh('git clean -qfdx')
+            Actions.sh('git clean -qfdx') unless params[:skip_clean]
             Helper.log.info 'Git repo was reset and cleaned back to a pristine state.'.green
           else
             paths.each do |path|
@@ -31,6 +31,7 @@ module Fastlane
       def self.details
         [
           "This action will reset your git repo to a clean state, discarding any uncommitted and untracked changes. Useful in case you need to revert the repo back to a clean state, e.g. after the fastlane run.",
+          "Untracked files like `.env` will also be deleted, unless `:skip_clean` is true.",
           "It's a pretty drastic action so it comes with a sort of safety latch. It will only proceed with the reset if either of these conditions are met:",
           "You have called the ensure_git_status_clean action prior to calling this action. This ensures that your repo started off in a clean state, so the only things that will get destroyed by this action are files that are created as a byproduct of the fastlane run."
         ].join(' ')
@@ -49,6 +50,12 @@ module Fastlane
           FastlaneCore::ConfigItem.new(key: :force,
                                        env_name: "FL_RESET_GIT_FORCE",
                                        description: "Skip verifying of previously clean state of repo. Only recommended in combination with `files` option",
+                                       is_string: false,
+                                       default_value: false),
+          FastlaneCore::ConfigItem.new(key: :skip_clean,
+                                       env_name: "FL_RESET_GIT_SKIP_CLEAN",
+                                       description: "Skip 'git clean' to avoid removing untracked files like `.env`. Optional, defaults to false",
+                                       optional: true,
                                        is_string: false,
                                        default_value: false)
         ]
